pyproject.toml
[project]
name = "anime1-desktop"
version = "0.1.0"
description = "Anime1 桌面应用 - 番剧浏览器"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.3.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pywebview==5.0",
    "pyinstaller>=6.0.0",
    "pillow>=10.0.0",
    "hanziconv>=0.3.2",
    "peewee>=3.17.0",
    "pyobjc-framework-WebKit>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
anime1 = "src.app:main"

[tool.setuptools]
packages = {find = {}}

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
